polish dist unitest
